--- src/containers/Leaderboard/LeaderboardBody.module.scss
@@ -1,54 +1,0 @@
-@import "~styles/variables.scss";
-
-.explorerBody {
-  padding: $xl-spacing;
-}
-
-.centeredBody {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-}
-
-.bodyHeading {
-  width: 100%;
-  display: flex;
-}
-
-.leaderList {
-  margin-top: $m-spacing;
-}
-
-.bountyListCentered {
-  flex-grow: 1;
-  display: flex;
-  align-items: center;
-}
-
-.centeredItem {
-  margin: 0 auto;
-}
-
-.bountyNumber {
-  margin-right: $s-spacing;
-}
-
-.sortGroup {
-  margin-left: auto;
-  display: flex;
-  align-items: flex-end;
-}
-
-.sortByText {
-  margin-right: $l-spacing;
-}
-
-.sortBy {
-  display: flex;
-  align-items: flex-end;
-}
-
-.loadMoreButton {
-  width: 100%;
-  text-align: center;
-}
