--- src/public-modules/Leaderboard/constants.js
@@ -1,2 +1,0 @@
-export const LIMIT = '10';
-export const PLATFORM = 'gitcoin,bounties-network';
