+++ src/public-modules/Leaderboard/index.js
@@ -3,13 +3,13 @@ const initialState = {
   loaded: false,
   error: false,
   leaderboard: {},
-  switchValue: 'issuer'
+  toggleValue: 'fulfiller'
 };
 
 const LOAD_LEADERBOARD = 'leaderboard/LOAD_LEADERBOARD';
 const LOAD_LEADERBOARD_SUCCESS = 'leaderboard/LOAD_LEADERBOARD_SUCCESS';
 const LOAD_LEADERBOARD_FAIL = 'leaderboard/LOAD_LEADERBOARD_FAIL';
-const USE_LEADERBOARD = 'leaderboard/USE_LEADERBOARD';
+const LEADERBOARD_TOGGLE = 'leaderboard/LEADERBOARD_TOGGLE';
 
 function loadLeaderboard() {
   return { type: LOAD_LEADERBOARD };
@@ -26,8 +26,8 @@ function loadLeaderboardFail(error) {
   return { type: LOAD_LEADERBOARD_FAIL, error };
 }
 
-function useLeaderboard(switchValue = 'Issuers') {
-  return { type: USE_LEADERBOARD, switchValue };
+function leaderboardToggle(toggleValue = initialState.toggleValue) {
+  return { type: LEADERBOARD_TOGGLE, toggleValue };
 }
 
 function LeaderboardReducer(state = initialState, action) {
@@ -59,11 +59,11 @@ function LeaderboardReducer(state = initialState, action) {
         error: true
       };
     }
-    case USE_LEADERBOARD: {
-      const { switchValue } = action;
+    case LEADERBOARD_TOGGLE: {
+      const { toggleValue } = action;
       return {
         ...state,
-        switchValue: switchValue.toLowerCase()
+        toggleValue: toggleValue.toLowerCase()
       };
     }
     default:
@@ -75,14 +75,14 @@ export const actions = {
   loadLeaderboard,
   loadLeaderboardSuccess,
   loadLeaderboardFail,
-  useLeaderboard
+  leaderboardToggle
 };
 
 export const actionTypes = {
   LOAD_LEADERBOARD,
   LOAD_LEADERBOARD_SUCCESS,
   LOAD_LEADERBOARD_FAIL,
-  USE_LEADERBOARD
+  LEADERBOARD_TOGGLE
 };
 
 export default LeaderboardReducer;
